+++ controller/konnect/constraints/constraints.go
@@ -21,6 +21,8 @@ type SupportedKonnectEntityType interface {
 		configurationv1alpha1.KongPluginBinding |
 		configurationv1alpha1.CredentialBasicAuth |
 		configurationv1alpha1.KongUpstream
+		configurationv1alpha1.KongUpstream |
+		configurationv1alpha1.KongCACertificate
 	// TODO: add other types
 
 	GetTypeName() string