+++ modules/manager/controller_setup.go
@@ -84,6 +84,8 @@ const (
 	KongCredentialsSecretControllerName = "KongCredentialSecret"
 	// KongCredentialBasicAuthControllerName is the name of the CredentialBasicAuth controller.
 	KongCredentialBasicAuthControllerName = "CredentialBasicAuth" //nolint:gosec
+	// KongCACertificateControllerName is the name of the KongCACertificate controller.
+	KongCACertificateControllerName = "KongCACertificate"
 )
 
 // SetupControllersShim runs SetupControllers and returns its result as a slice of the map values.
@@ -388,6 +390,15 @@ func SetupControllers(mgr manager.Manager, c *Config) (map[string]ControllerDef,
 					c.DevelopmentMode,
 					mgr.GetClient(),
 					konnect.WithKonnectEntitySyncPeriod[configurationv1alpha1.KongUpstream](c.KonnectSyncPeriod),
+				),
+			},
+			KongCACertificateControllerName: {
+				Enabled: c.KonnectControllersEnabled,
+				Controller: konnect.NewKonnectEntityReconciler[configurationv1alpha1.KongCACertificate](
+					sdkFactory,
+					c.DevelopmentMode,
+					mgr.GetClient(),
+					konnect.WithKonnectEntitySyncPeriod[configurationv1alpha1.KongCACertificate](c.KonnectSyncPeriod),
 				),
 			},
 			KongPluginBindingControllerName: {