+++ CHANGELOG.md
@@ -52,6 +52,8 @@
   [#506](https://github.com/Kong/gateway-operator/pull/506)
 - Add `KongConsumerGroup` reconciler for Konnect control planes.
   [#510](https://github.com/Kong/gateway-operator/pull/510)
+- Add `KongCACertificate` reconciler for Konnect control planes.
+  [#626](https://github.com/Kong/gateway-operator/pull/626)
 - Added command line flags to configure the certificate generator job's images.
   [#516](https://github.com/Kong/gateway-operator/pull/516)
 - Add `KongPluginBinding` reconciler for Konnect Plugins.